+++ tests/e2e/module/module.tests.ps1
@@ -136,3 +136,41 @@ Describe 'Style rules' {
         }
     }
 }
+
+#Code Count
+Describe 'Code Statistics' {
+
+        $files = @(
+            Get-ChildItem "$root" -Include *ps1, *psm1
+            Get-ChildItem "$($root)functions" -Include *.ps1, *.psm1 -Recurse
+            Get-ChildItem "$($root)\tests" -Include *.ps1, *.psm1 -Recurse
+        )
+        $lineCount = 0
+        $NewlineCount = 0
+        $realline = 0
+        foreach ($file in $files) {
+            $lines = [System.IO.File]::ReadAllLines($file.FullName)
+            $lineCount += $lines.Count
+
+            for($i = 0; $i -lt $($lines.Count); $i++) {
+                if ($lines[$i] -match '^\s*$') {
+                    $NewlineCount ++
+                }
+                else{
+                    $realline ++
+                }
+            }
+        }
+
+    It "$($moduleName) Code base has $($lineCount) total lines" {
+        $true | Should Be $true
+    }
+
+    It "$($moduleName) Code base has $($realline) code lines" {
+        $true | Should Be $true
+    }
+
+    It "$($moduleName) Code base has $($NewlineCount) blank lines (Style Lines)" {
+        $true | Should Be $true
+    }
+}