+++ tests/e2e/reports/Export-CombinedReports.tests.ps1
@@ -9,11 +9,14 @@ Describe "Export-CombinedReports PS: $PSVersion" {
 
     Setup -Dir nofiles
     Setup -Dir results
+    Setup -Dir single
+    Setup -Dir resultsxlsx
 
-    Copy-Item "$here\sample.ckl" "TestDrive:\sample.ckl"
-    Copy-Item "$here\sampleV2.ckl" "TestDrive:\sampleV2.ckl"
-    Copy-Item "$here\Nessus_Sample_Linux.nessus" "TestDrive:\Nessus_Sample_Linux.nessus"
-    Copy-Item "$here\Nessus_Sample_Windows.nessus" "TestDrive:\Nessus_Sample_Windows.nessus"
+    Copy-Item "$Global:testData\CKL\CKLv1\sample.ckl" "TestDrive:\sample.ckl"
+    Copy-Item "$Global:testData\CKL\CKLv1\sample.ckl" "TestDrive:\single\sample.ckl"
+    Copy-Item "$Global:testData\CKL\CKLv2\sampleV2.ckl" "TestDrive:\sampleV2.ckl"
+    Copy-Item "$Global:testData\Nessus_Scans\Nessus_Sample_Linux.nessus" "TestDrive:\Nessus_Sample_Linux.nessus"
+    Copy-Item "$Global:testData\Nessus_Scans\Nessus_Sample_Windows.nessus" "TestDrive:\Nessus_Sample_Windows.nessus"
 
     Context "Strict mode" {
 
@@ -43,8 +46,23 @@ Describe "Export-CombinedReports PS: $PSVersion" {
             {Export-CombinedReports -NESSUS "TestDrive:\nofiles" -Output "TestDrive:\"} | Should Throw "No Nessus Files Found"
         }
 
-        It "Should Throw ' CKL file failed to process'" {
-            {Export-CombinedReports -CKLFILES "TestDrive:\" -Output "TestDrive:\results"} | Should Throw "$($(Get-ChildItem -Path "TestDrive:\sampleV2.ckl").name)  CKL file failed to process"
+        It "Should Throw 'CKL file failed to process'" {
+            {Export-CombinedReports -CKLFILES "TestDrive:\" -Output "TestDrive:\results"} | Should Throw "$($(Get-ChildItem -Path "TestDrive:\sampleV2.ckl").name) CKL file failed to process"
+        }
+
+        It "Should Create 2 (csv) files" {
+            Remove-Item -Path "TestDrive:\sampleV2.ckl"
+            Export-CombinedReports -CKLFILES $TestDrive -NESSUS "$TestDrive" -Output "TestDrive:\results" -name "Sample"
+            $reportfile = Get-ChildItem -Path "TestDrive:\results"
+            $reportfile.count | Should Be 2
+            $reportfile[0].extension | should be ".csv"
+        }
+
+        It "Should Create 2 (xlsx) files" {
+            Export-CombinedReports -CKLFILES $TestDrive -NESSUS "$TestDrive" -Output "TestDrive:\resultsxlsx" -name "Sample" -xlsx
+            $reportfile = Get-ChildItem -Path "TestDrive:\resultsxlsx"
+            $reportfile.count | Should Be 2
+            $reportfile[0].extension | should be ".xlsx"
         }
 
     }