+++ tests/run-tests.ps1
@@ -1,5 +1,6 @@
 ﻿$moduleName = "PSIASTAND"
 $here = Split-Path -Parent $MyInvocation.MyCommand.Path
+$Global:testData = "$here\data"
 Import-Module "$here\..\$($moduleName)"
 Invoke-Pester
 Remove-Module $moduleName