+++ tests/unit/stigviewer/main/import-ckl.tests.ps1
@@ -7,8 +7,8 @@ $PSVersion = $PSVersionTable.PSVersion.Major
 
 Describe "Import-CKL PS: $PSVersion" {
 
-    Copy-Item "$here\sample.ckl" "TestDrive:\sample.ckl"
-    Copy-Item "$here\sampleV2.ckl" "TestDrive:\sampleV2.ckl"
+    Copy-Item "$Global:testData\CKL\CKLv1\sample.ckl" "TestDrive:\sample.ckl"
+    Copy-Item "$Global:testData\CKL\CKLv2\sampleV2.ckl" "TestDrive:\sampleV2.ckl"
 
     Context "Strict mode" {